requirements.txt
streamlit>=1.32.2
requests>=2.31.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=1.0.0
python-dotenv>=1.0.1
Pillow>=9.0.0